Working transaction endpoints

+++ src/models/account.model.ts
@@ -2,7 +2,7 @@ import { DataTypes, Sequelize } from "sequelize";
 
 module.exports = (sequelize: Sequelize) => {
     const Account = sequelize.define("Account", {
-        capital: DataTypes.INTEGER
+        capital: DataTypes.FLOAT
     });
   
     return Account;
