+++ src/models/transaction.model.ts
@@ -2,7 +2,7 @@ import { DataTypes, Sequelize } from "sequelize";
 
 module.exports = (sequelize: Sequelize) => {
     const Transaction = sequelize.define("Transaction", {
-        amount: DataTypes.INTEGER,
+        amount: DataTypes.FLOAT,
         date: {
             type: DataTypes.DATE,
             defaultValue: DataTypes.NOW
